One working test so far.
Just "finished" linesPerCycle, but there are several problems with the
test. The most important being that I believe EasyUO is case
insensitive. This means that linesPerCycle needs to disregard case.
Also, the expression that is passed to it is currently parsing as
'string', which I believe is not good because I added an integer and I
am not sure if I want to have those as two logically different things
since it does not appear to matter to EasyUO. The given subroutine that
is called may complain however.